==> scripts/01-sample_list.R <==
rm(list = ls())
(myfiles <- list.files(pattern = "_R1.fastq.gz"))
(samples <- gsub("_R1.fastq.gz", "", myfiles))
write.table(samples, file = "samples", quote = FALSE, row.names = FALSE, col.names = FALSE)
